moved SegmentColorUserShaderManager into segment_color.ts
